move all files connected to App into App folder
